backend/payment/requirements.txt
stripe==7.3.0
pydantic==2.5.0
requests==2.31.0
psycopg2-binary==2.9.7